+++ src/Simplexcel/XlsxInternal/XlsxCell.cs
@@ -2,7 +2,7 @@
 
 /// <summary>
 /// A Cell inside the sheet.xml file
-/// ECMA-376, 3rd Edition, Part 1, 18.3.1.4 c (Cell)
+/// ECMA-376, 5th Edition, Part 1, 18.3.1.4 c (Cell)
 /// </summary>
 internal class XlsxCell
 {
@@ -21,7 +21,15 @@ internal class XlsxCell
     internal int StyleIndex { get; set; }
 
     /// <summary>
+    /// The f element
+    /// </summary>
+    internal string Formula { get; set; }
+
+    /// <summary>
     /// The v element
     /// </summary>
+    /// <remarks>
+    /// For formula cells, contains the most recent computed value as specified in ECMA-376, 5th Edition, Part 1, 18.17.6.6 Value Representation
+    /// </remarks>
     internal object Value { get; set; }
 }